requirements.txt
django==6.0.1
django-cors-headers==4.9.0
djangorestframework==3.16.1
